MINOR: mux-h1/htx: Set flag to stay on same server on the conn_stream
With the legacy HTTP representation, this preference is handled by the HTTP
transaction by setting the flag TX_PREFER_LAST when it is cleaned up, for
keep-alive connections. But, with the HTX representation, keep-alive connections
are handled by the mux h1 and a new stream is created for each connection. So
there is no way to get flags for a transaction using flags of the previous
one. Instead, to set this preference on the HTTP transaction, we set it on the
conn_stream. To do so, the flag CS_FL_PREFER_LAST has been added. The mux h1 set
it on the CS when it is created, depending on information of the previous
transaction (in this case, its status code). Then, this flag is tested when a
new server is assigned to the stream.