MINOR: http: Use CS_FL_PREFER_LAST instead of TX_PREFER_LAST
The flag CS_FL_PREFER_LAST has been added on the conn_stream because it was
impossible to rely on the txn for the HTX representation. It is now easier to
also use this flag for the legacy representation. So the flag TX_PREFER_LAST has
been removed and the flag CS_FL_PREFER_LAST is used in all cases.