REORG: includes: move ist.h from common/ to import/
Fortunately that file wasn't made dependent upon haproxy since it was
integrated, better isolate it before it's too late. Its dependency on
base.h was the result of the change from config.h, which in turn wasn't
correct. It was changed back to inttypes.h for size_t. The recently
added reference to MAX() was changed as it was placed only to avoid a
zero length in the non-free-standing version and was causing a build
warning in the hpack encoder.